# Startup `flutter` faster (Only access globals.deviceManager if actually setting something)