--- a/packages/flutter_tools/lib/src/runner/flutter_command_runner.dart
+++ b/packages/flutter_tools/lib/src/runner/flutter_command_runner.dart
@@ -256,7 +256,10 @@ class FlutterCommandRunner extends CommandRunner<void> {
         }
 
         // See if the user specified a specific device.
-        globals.deviceManager?.specifiedDeviceId = topLevelResults['device-id'] as String?;
+        final String? specifiedDeviceId = topLevelResults['device-id'] as String?;
+        if (specifiedDeviceId != null) {
+          globals.deviceManager?.specifiedDeviceId = specifiedDeviceId;
+        }
 
         if ((topLevelResults['version'] as bool?) ?? false) {
           globals.flutterUsage.sendCommand('version');