@jonahwilliams 2022-09-13
This seems like a reasonable change, but is something that may accidentally regress if not tested in some way. Perhaps we could cover this with a CI benchmark that measures start up time of the flutter tool?

@jensjoha if you'd like I'd be more than happy to walk you though adding one. since you're doing so much work here and elsewhere to improve performance I think it would be great to track it