@jensjoha 2022-09-14
I agree that a benchmark would be good.
I had originally wanted one for https://github.com/flutter/flutter/pull/107395 as well, but couldn't seem to find anything about how to add one, so if you can shed some light on that, that would be great!