@jonahwilliams 2022-09-14
I think the best example to work from would be something like the analyzer benchmark:

https://github.com/flutter/flutter/blob/195a1cc413bde5f7c1d194204608c4bc20659124/dev/devicelab/lib/tasks/analysis.dart#L23-L57

In this we simply run some command and record the total time. In this case, maybe just `flutter -h`. So there are a few parts to this:

1. Add a function structured like the analyzer task function above.
2. Add 3 files in dev/devicelab/bin/tasks.dart, one for each host platform variant. You could name them something like flutter_tool_startup__windows, flutter_tool_startup__macos, et cetera. They can all have almost the same content and just wrap the call to the task function in a `main`.
3. Add 3 new entries to the ci.yaml file https://github.com/flutter/flutter/blob/master/.ci.yaml that refers to the tasks you added above. I think Linux_android analyzer_benchmark is a good example to copy from. All of these should be marked `bringup: true`.
4. Add entries to TESTOWNERS for each test with yourself + flutter/tool as an owner.

If you land this before any of these PRs, then we can observe the improvements too